requirements.txt
fastapi
uvicorn
python-multipart
boto3
openai
PyMuPDF
python-dotenv
streamlit
requests
